+++ src/main/java/com/xinao/sync/entity/xinao/Sale.java
@@ -1,6 +1,5 @@
 package com.xinao.sync.entity.xinao;
 
-    import java.math.BigDecimal;
     import com.baomidou.mybatisplus.extension.activerecord.Model;
     import java.time.LocalDateTime;
     import java.io.Serializable;
@@ -25,11 +24,11 @@ package com.xinao.sync.entity.xinao;
 
     private Integer id;
 
-    private BigDecimal price;
+    private Double price;
 
     private Integer amount;
 
-    private BigDecimal cost;
+    private Double cost;
 
     private LocalDateTime buyDate;
 
@@ -49,17 +48,17 @@ package com.xinao.sync.entity.xinao;
 
     private String flowid;
 
-    private BigDecimal mprice;
+    private Double mprice;
 
     private Integer mamount;
 
-    private BigDecimal mcost;
+    private Double mcost;
 
-    private BigDecimal hprice;
+    private Double hprice;
 
     private Integer hamount;
 
-    private BigDecimal hcost;
+    private Double hcost;
 
 
     @Override
